+++ compiler/libchplcomp/include/chpl/AST/BaseAST.h
@@ -3,8 +3,8 @@
 
 #include "chpl/AST/ASTTag.h"
 #include "chpl/AST/ID.h"
-#include "chpl/AST/memory.h"
 #include "chpl/AST/ASTTypes.h"
+#include "chpl/Util/memory.h"
 
 namespace chpl {
 namespace ast {
